requirements.txt
openai==0.28.0
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.39.0
